Collapsing layout & configuration
Added a configuration option to show completed quests.

Quests that have an unknown state as of now will appear with the Gray
text color. Eg: Tutorial Island, Natural history quiz, ...

Configuration also allows the user to change colors for completed and
in-progress quests.

The collapsing layout allows for dynamic resizing by toggling the
visiblity of elements as the updates are now occuring on game tick. This
should be better than removing all the elements and then adding them
back again.