Joining Clan Chat AssertionError
When joining Clan Chats the ConfigChanged event would get called and
cause an AssertionError. This would cause the client to crash during the
updateQuestList method during quest.getState(c).

Debugging displayed the Client as not-null and if checking the Client
resulted in a not-null instance but would still throw the error. For now
the easiest way around it is to remove the ConfigChanged event and I
will be doing so for now.