Version Bumps, 1.7 Fixes
Fixed issue with IconTextField.addKeyListener causing the plugin to
crash. I'm not a super big fan of the implementation I have here but for
the time being it will hold over. I'll take a look at it again sometime
during the week.

Thanks to iFester for the heads up about the plugin not showing up.