Update name and link for Rag and Bone Man I
Currently, the wiki auto-redirects the previous link to the correct URL. This is the new URL on the wiki and reflects the name in Runescape. 

Because of this name change to "Rag and Bone Man I", the plugin fails to identify if the user has completed the quest. Updating the name fixes the bug