Panel styling / setup + Skill Icons
This should be fairly close to what I want the plugin to look like. I
have to accept that there are like 3 pixels that I cannot control inside
of a DynamicGridLayout however it is likely that I just don't know what
I am doing.

I might merge the Requirement & Reward Panels into one since they are
likely going to look very similar.

Need to add in config settings and we should be good to go for adding
the event handlers and updating.